requirements.txt
# Core dependencies
grpcio==1.60.0
grpcio-tools==1.60.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.1
celery==5.3.4

# Machine Learning
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.2
tensorflow==2.15.0
shap==0.43.0
lime==0.2.0.1

# Security and Encryption
cryptography==41.0.8
pyjwt==2.8.0
bcrypt==4.1.2
argon2-cffi==23.1.0

# Event Sourcing and Messaging
kafka-python==2.0.2
pydantic-settings==2.1.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-benchmark==4.0.0
hypothesis==6.92.1
faker==20.1.0

# Development and DevOps
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0
docker==6.1.3
kubernetes==28.1.0

# Monitoring and Logging
prometheus-client==0.19.0
structlog==23.2.0

# Utilities
click==8.1.7
rich==13.7.0
typer==0.9.0
